compute: feed MV sink frontier from persist write handle
This commit changes the way the MV sink produces updates to the shared
sink frontier. Previously it would inspect the `persist` stream and
forward that stream's frontier. The issue with this approach is that, in
the face of ingestion spikes, the `persist` stream could fall behind,
which would also make the MV sink's reported write frontier fall behind.

This lead to a regression in the `CrossJoin` feature benchmark. In this
benchmark, a large constant MV is created. As soon as the MV dataflow
has produced its snapshot and has reported the empty frontier, the
controller instructs the replica to drop it. If the reporting of the
empty frontier is delayed because the MV sink is busy reading back the
entire snapshot from persist, the controller's drop command also gets
delayed. The result is that queries against the cluster get slowed down
by the MV dataflow's processing, more than they would have been had the
MV dataflow been dropped quickly.

Attaching the shared write frontier to the `WriteHandle`-based frontier
stream used by the `mint` operator avoids the issue.