durable-cache: Fix partial application of times
Previously, when syncing with the persist shard, the durable cache
would apply each batch from `ListenEvent::Updates` to a map in memory
as it received them. While applying the updates, the durable cache
would assert that each retraction had a matching addition and that
each addition didn't already exist in the cache.

Subscribe does not guarantee that a single `ListenEvent::Updates` will
contain all the updates of a single timestamp or that the updates
within a single timestamp will arrive in any particular order. This
could cause the durable cache to incorrectly panic thinking there's an
invalid update.

For example, consider the following sequence of updates:

  1. `ListenEvent::Updates([((k, v), ts, -1)])`
  2. `ListenEvent::Updates([((k, v), ts, +1)])`

While processing (1), the durable cache would think that the retraction
doesn't have a matching addition and panic.

For another example, consider the following sequence of updates:

  1.  `ListenEvent::Updates([((k, v1), ts, +1), ((k, v2), ts, +1)])`
  1.  `ListenEvent::Updates([((k, v1), ts, -1)])`

While processing (1), the durable cache would think that the addition
of `v2` is invalid because `v1` was never retracted and panic.

This commit fixes this issue by batching the contents of
`ListenEvent::Updates` in a map, and only applying them once a
`ListenEvent::Progress` arrives for the batched timestamps. This
ensures that we always process all the updates for a timestamp at once.

Fixes #MaterializeInc/database-issues/issues/8799