delete legacy upsert error handling
This PR removes code that was added during incident 72 (a long time
ago). The incident was caused because we introduced a backward
incompatible proto serialization of certain upsert errors. We then added
the code removed in this PR that both handled the incompatibility and
also issued corrections so that the correct serialization ends up in the
shards.

By now all such legacy error instances have been compacted away and so
this code is not needed anymore.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>