storage/pg: don't rely on owned capabilities
This PR fixes a long-standing bug of postgres sources that was a result
of using owned capabilities in an async operator. Fallible async
operators use the standard Rust `?` operator to encode erroring control
flow. Unfortunately this doesn't play well with the rest of the timely
integration since when `?` is executed on an error the entire stack of
local variables will be destructed. If that stack contains capabilities
then timely will be notified about frontier updates that are incorrect,
and if the timing is right it might have enough to persist those
incorrect results to persist.

As a workaround the API for fallible async operators passes the initial
capabilites through a reference to operator implementation with the
expectation that the operator only uses those capabilties to communicate
progress. This PR implements exactly this idea for pg rewinds.

Fixes: MaterializeInc/database-issues#7008
Fixes: MaterializeInc/database-issues#8803

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>