storage: make CDCv2 ingestions not depend on progress collections
CDCv2 collections are in their own timeline, not in the
EpochMilliseconds timeline, and have no need for the reclocking
information it contains. It should not be necessary to hold back the
read frontier of the progress collection for CDCv2 exports, and it would
be wrong to initialize the `since` of CDCv2 exports to the `since` of
their progress collection.