storage controller: reveal the concrete StorageCollections impl
The `StorageCollections` trait is meant to expose a simple interface to
components outside of storage that allows them to monitor frontiers and
such.

Naturally, there is a lot of overlap and communication between the data
held in the `StorageCollectionsImpl` and the `StorageController`, but
because the latter was only having access to the former via a trait
object it necessitated the trait to have many more methods and surface
area just so that the two could cooperate.

In reality, the storage controller and storage collections are tightly
coupled and putting a trait between them is adding unnecessary
complexity.

This PR specializes the type of `StorageCollections` in the storage
controller. This allows the storage controller to call non-exposed
methods, non-public methods on the `StorageCollectionsImpl` value and
allows to drastically reduce the surface area offered by the trait. So
much so that its implementation can could be replaced by just some
shared pieces of data over an `Arc<Mutex<_>>`.

This PR is only code movement and deletion. Future PRs will build on
this new structure by consolidating and simplifying collection handling.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>