source/pg: detect replication slot overcompaction
The source would previously select a resumption point based on the
resumption points of its outputs (which is correct) advanced by the
resumption point available upstream (which is incorrect).

The reason is that if the upstream slot gets dropped and recreated or
somehow compacts further that it should we lose the ability to read
updates at the LSN offsets we need. When that happens it is crucial to
detect that and put the source in an errored state, otherwise we risk
presenting incorrect to the user.

This PR tightens our reasoning for choosing the expected resumption
point such that the advancement that would previously be done
unconditionally is only performed for outputs that have not been
ingested yet.

Fixes MaterializeInc/database-issues#8669

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>