storage: retain counter statistics between epochs
Storage dataflows (computing sources and sinks) may be recreated for
various reasons (ingestion errors, addition of new source exports, ALTER
SINK, ...). When a storage dataflow is recreated, its statistics epoch
is increased, as a means of differentiating statistics collected for the
old dataflow instance from statistics collected for the new dataflow
instance.

Prior to this change, increasing the epoch would reset all collected
statistics and cause all in-flight statistics with the old epoch to be
ignored. For gauge statistics that report a current point-in-time value
this is the correct behavior as values emitted by the old dataflow might
be wrong for the new dataflow. However, for counter statistics that keep
track of the number of events of a certain type that have occurred so
far, the behavior is less desirable. After all, the counted event did
occur and not reporting it just because it raced with a dataflow restart
can lead to confusing results. For example, we might report zero
`updates_committed` for a source that clearly contains updates, because
a new table was added in the time between when the updates were
committed and when their statistics could be reported.

The semantics of our counter statistics are weak enough to allow this
kind of undercounting, but it is still annoying, both from a UX
perspective and for writing tests that try to ensure the statistics
collection is working as expected.

This commit attempts to improve things by ensuring that counter
statistics are retained through epoch changes, rather than discarded.
Gauge statistics are still cleared/reset as before.