compute/mv-sink: force consolidation after the snapshot
This commit makes the MV sink always perform a consolidation of (the
relevant parts of) the correction buffer after it has received all
updates for the snapshot, both from `desired` and `persist`. We know
that these updates will cancel out, so we have an opportunity to shed a
large amount of memory by forcing a consolidation at this time. This is
especially important in read-only mode when the MV sink does not produce
any batches, which is the process that normally forces consolidation to
happen.