Fix self-hosted pgwire cancellation
Fix self-hosted pgwire cancellation.

We previously did not sent cancellation requests if we didn't have a
directory containing mappings from suffixes to DNS addresses.

We now require that the cancellation directory is provided or that we
are using a static resolver, and we also use that same static resolver
for cancellation requests.