materialized: bundle console in image
To make it even easier to get started with Materialize, this commit
bundles a recent copy of the console into the emulator image
(i.e., the materialize/materialized image).

This reuses the console packaging that we already have in place to
bundle the console with the self-managed distribution.

We'll need to remember to periodically bump the console image, but it
shouldn't be too bad--the only time we *must* bump the image is when we
make a backwards incompatible change to a feature the console is
depending on, which doesn't happen very often.