source/kafka: pull out reader render function
This commit pulls out the render code for the Kafka source reader
operator into a separate function, in anticipation of the addition of a
second metadata fetcher operator.

This change is almost exclusively code movement. It gets its own commit
to reduce the diff of the next commit that separates the metadata
fetching logic.