source/kafka: split out a metadata fetcher
This commit pulls out the metadata fetching code from the reader
operator and moves it into a separate metadata fetcher operator. The
latter is a single-worker operator that broadcasts its output to all
workers. The intent of this change is to share metadata updates between
workers, rather than duplicating the fetching work on each worker.