storage: infrastructure for independent source output streams
This PR puts in the basis for allowing a source implementation to
produce an independent DD collection per output.

The situation before this PR was that the `SourceRender` trait required
a single, multiplexed DD collection to be produced, of type `(usize,
D)`, where the `usize` designated the output. Since all outputs were
multiplexed, a single frontier had to describe their overall progress,
which described the "slowest" one. This is generally fine when all
outputs more or less march forwards together but it's not ok when a new
subsource is added to a source that has otherwise been running for a
while. In this situation the upper frontier of the multiplexed
collection would necessarily have to stay stuck until the new subsource
finished its snapshot and caught up with the other ones, making the
previously healthy sources unavailable for all this time.

This PR fixes this by requiring a `BTreeMap<GlobalId, Collection>`
output type from source implementations. This way each subsource can be
driven independently and a new subsource can be added without imposing a
frontier stall to the previously ingested subsources.

This PR does not change any of the source implementations to take
advantage of this new interface, since that would create a giant PR.
Instead, this only changes the interface and the fallout of all the
changes in the various generic parts of the pipeline. Follow up PRs will
be done that target individual source implementation and change them to
be directly produce multiple collections.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>