storage: fix reporting of empty sink frontiers
When a sink dataflow is installed, it can happen that the `resume_upper`
is determined to be the empty frontier, e.g. when the sink has
previously produced all its output already. In this case we immediately
shut down the sink dataflow. Prior to this change, we'd do so without
updating the shared sink frontier, so neither the `StorageState` nor the
controller would learn about the fact that the sink dataflow has shut
down. This is fixed in this commit.