kafka-source: detect topic deletion in metadata fetcher
This commit moves detection of topic recreation from the reader operator
into the metadata fetcher. This is done to make the metadata fetcher
aware of the unrecoverable error so it can seal the collection by
emitting a probe with an empty upstream frontier. Prior to this change,
the collection wouldn't be sealed under the reclock-to-latest scheme,
and a replica restart could cause it to receive new, incorrect updates.

Apart from fixing the sealing bug, this commit makes two functional
changes:

 * It adds detection of topic deletion. That was easy to add and not
   having it seemed like an oversight.
 * It omits the "high watermark regressed"/"partition count regressed"
   detail in the "topic recreated" error. It didn't feel like a useful
   piece of information and omitting it simplifies the code a lot.