envd: make tests robust against slower read hold downgrade
The changes made in the previous commits make it so the compute
controller is slower at applying read hold downgrades. In tests that
spawn two envd instances in sequence that can lead to issues because the
first instance might still try to apply read hold downgrades while the
second instance is already running, causing `PersistEpoch` conflicts.
This race condition existed before but apparently wasn't hit because
read hold downgrades by the first instance didn't get delayed long
enough.

The fix chosen here is to convert the tests that want to run two envd
instances in sequence from using `TestHarness::start` to using
`TestHarness::start_blocking`. The latter runs the `TestServer` in a
separate tokio runtime that gets dropped when the `TestServer` is
dropped, ensuring that no controller tasks can still be alive when the
second envd instance is started.