storage: fix shutdown token for copy to s3 sink
The COPY TO S3 sink must not flush files to S3 during shutdown, as those
files may only be partially written.

This can cause correctness issues when running copy operations on
multiprocess replicas, as the lagging replica's copy sink will get
dropped partway through.

The solution is to properly convert the buttons returned by the async
operators in the S3 sink's implementation into tokens that are dropped
when the sink is dropped. This ensures that the oneshot sink is properly
shut down when dropped, before it can write partial data to S3.

Fix MaterializeInc/database-issues#8599.