storage/sources/postgres: avoid metadata queries that are expected to fail
When fetching schema info from PostgreSQL, the PostgreSQL source
currently issues a query that's known to be compatible only with recent
versions of PostgreSQL.  If the query fails, the source retries with a
version of the query that's compatible with older versions of
PostgreSQL.

While this works fine for customers running versions of PostgreSQL
before v15, it has an annoying side effect: those customers can observe
the failed queries in their PostgreSQL log files. This has led to
confusion and complaints from at least two customers.

This commit changes the PostgreSQL source to do eager server version
sniffing and issue a query that's known to be compatible with the
announced server version. This should avoid leaving failing queries in
the upstream PostgreSQL server's logs, and hopefully avoids future
customer confusion.