adapter: emit existence notice in CREATE CONNECTION IF NOT EXISTS
If the named connection already exists when running a CREATE CONNECTION
IF NOT EXISTS command, emit a notice indicating the existing connection.

This brings CREATE CONNECTION IF NOT EXISTS into alignment with the
behavior of all other CREATE ... IF NOT EXISTS commands, which all emit
notices. Somehow connections were overlooked.