kafka: don't shutdown metadata fetcher on error
Shutting down the metadata fetcher operator releases its capability on
the probe output. When the reclock-to-latest reclocking is enabled this
makes the remap operator believe that the upstream has advanced to the
empty frontier, making it seal the collection. We must thus not shut
down the metadata fetcher when it encounters a connection error, as the
source did not in fact advance to the empty frontier in this case.