test: remove manual setting of Kafka metadata interval
A bunch of tests set the `kafka_default_metadata_fetch_interval` or the
source `TOPIC METADATA REFRESH INTERVAL` to low values. This was done
because the default for `kafka_default_metadata_fetch_interval` was 30s
and some tests were slowed down by that. The default has now been
changed to 1s, so manually lowering the fetch interval is not necessary
anymore.

Removing the manual setting of the Kafka metadata interval in tests gets
rid of some noise. But more importantly, it fixes the issue of some
tests setting the interval to a very low value, like '10ms', which now
that the metadata fetch interval influences the source tick frequency
could negatively impact performance and make tests slower or even time
out.