storage: in upsert StateValue allow converting finalized tombstones
Before, we were using the wrong assumption that the new feedback UPSERT
operator could not yield finalized tombstone values, because it only
writes finalized values that "exist". But it can happen that
`ensure_decoded` puts in place a finalized tombstone when state values
consolidate to a diff of zero.

We say "can happen" above, because `merge_update` returns whether or not
a merged value can be deleted, and this returns true when the diff_sum
is zero. So in most cases the finalized tombstone would be cleaned up
right away. However, not all callers immediately (or at all) act on this
"can delete" information, which leads to finalized tombstones actually
existing in state.