storage: in upsert, determine readiness based on outer ts
This is not a fix for a bug that we observed, but I think it's still
incorrect to check using these combined `(ts, subtime)` timestamps.

I believe there is a general problem with the subtime approach, in a way
the outside code is "lying" to the operators in side the subtime scope.
A pattern in timely operators is to stash updates of a given timestamp
`t` until we know that we have seen all updates for that timestamp, at
which time we process them all together. The way operators usually check
that timestamp `t` is ready for processing is something like
`!input_frontier.less_equal(t)`.

Within the Subtime scope this is broken. For example, say we have
updates at timestamp `(5, Subtime(0))`. Now the input frontier advances
to `(5, Subtime(1))`. The operator will now naively assume that updates
at time `(5, Subtime(0))` are ready for processing, but in fact we
haven't yet seen all the updates of the "full" outer timestamp `5`.

It feels incorrect that the operator has to know about this special
timestamp and have special "unwrapping" code that needs to be applied
before checking for readiness.