storage: in feedback upsert, fix handling of existing state values
Before, while working off the "upsert commands" (aka. input) we were
taking the existing value out of `command_state` when processing and we
were only putting in a new value when processing commands with
`DrainStyle::AtTime`. This is problematic when processing commands of
multiple timestamps in one invocation of `drain_staged_input`:
processing the first update for a key `k` at timestamp `t` takes away
any value that we might have retrieved from state, then when processing
another update for key `k` at timestamp `t+1`, we think there is no
previous value and don't emit a retraction. Meaning we suddely have
multiple values for the same key in our collection, which is not legal.

Now, we just don't take away the state value without putting something
back.

The astute reader might now wonder how this can be correct. Say our
upsert state contains:

```
k1 -> v2
```

The input frontier is at at a time that allows processing and the
persist input frontier is at `[11]`.

We're processing these upsert commands, tuples of `(key, timestamp,
value)`:

```
(k1, 10, v2)
(k1, 11, v3)
```

Naively (and somewhat correctly), you would assume the correct updates
to emit when processing this, one timestamp after another, would be (now
additionally with diffs):

```
((k1, 10, v2), -1)
((k1, 10, v2), 1)
((k1, 11, v2), -1)
((k1, 11, v3), 1)
```
(Those first two updates might seem nonsensical, but the code doesn't
actually check to see if the values differ, it simply retracts and
updates.)

The answer lies in the fact that we're _only_ allowed to process updates
with timestamp `11` when they are not beyond the persist frontier
anymore, that is the frontier is at say `11`. Meaning what we have in
state is the _global view_ of upsert state as of that persist frontier,
meaning it is correct to emit updates at `11`. And anything we emit at
"lower" timestamps (timestamps that are not beyond the persist frontier,
for sticklers) will not be written down because the output shards upper
is already past that. All that will be written down is:

```
((k1, 11, v2), -1)
((k1, 11, v3), 1)
```

Which is the correct updates to write down given the current contents of
the shard at upper `11`. This _is_ subtle, and I'm note sure I like it,
but I also think it's correct.

As a follow-up, we could say that we don't process updates that are not
beyond the persist frontier, because we know that they won't be written
down anymore. But this is not needed to fix this bug.