storage: Correct output index for source exports
Previously, the output index of each ingestion export was calculated
assuming that index 0 was reserved for the primary source relation.
When `force_source_table_syntax` is enabled, the primary source
relation is not included in the exports, and therefore caused all
export indexes to be off by one.

This commit fixes the issue by decrementing all export indexes by 1 if
the primary source relation was not included in the exports.

It's unclear to the author how errors are being handled when the
primary relation is not included in the exports.

Works towards resolving #MaterializeInc/database-issues/issues/8620