New structured-only blob variant
This involves making `fn records` return an option, which has many
knock-on consequences. I've punted on a few of these, specifically for
codepaths where we know that we will not yet encounter that case in
practice. The expectation is that this code will either be removed
entirely or modified in some near-future followups.