storage: Simplify dataflow as_of computation (#30972)
Previously, async storage workers would calculate the as_of of a
dataflow from the since of the dataflow's output's remap shards. If
there was more than one distinct remap shard among the outputs, then the
storage worker would panic. It's expected that the only collection that
will ever have a remap shard is the ingestion collection itself.

Furthermore, we are planning to remove the ingestion collection from the
outputs of the dataflow (in fact there's already a feature flag that
does this). If the ingestion is removed from the outputs, then no output
will have a remap shard, and the as_of will always be empty.

This commit simplifies the existing as_of calculation and fixes the
as_of calculation when the ingestion collection is removed from the
outputs. It does this by calculating the as_of directly from the
ingestion's remap shard. Additionally, it asserts that if any of the
outputs have a remap shard, then it must be equal to the ingestion's
remap shard.

Works towards resolving #MaterializeInc/database-issues/issues/8620