storage: make KAFKA_METADATA_FETCH_INTERVAL a dyncfg
This commit removes the `KAFKA_DEFAULT_METADATA_FETCH_INTERVAL` legacy
config and replaces it with a `KAFKA_METADATA_FETCH_INTERVAL` dyncfg.
Like `{PG,MYSQL}_OFFSET_KNOWN_INTERVAL`, this makes it possible to
dynamically observe updates to this interval through a `ConfigSet`,
apart from removing some boilerplate.

The flag also loses its "default" part and is now the only thing
determining the probe ticker interval. This is the first step in
removing the user-visible option to configure the Kafka source metadata
fetch interval and is done here because it simplifies dynamically
updating the probe ticker interval in the next commit.