storage: make probe intervals runtime configurable
This commit makes it possible to change the tick interval of the
upstream frontier probes without a source restart. To this end, the
`probe::Ticker` receives a closure with which it can request the current
interval, which it calls between ticks. The various source
implementations implement the closure to read their respective dyncfgs.