Consolidate produces batch of data
Instead of flattening the outputs of consolidate, provide the output as a
whole batch, encoded as an `Vec<C>`. This allows downstream operators
to act on a whole batch at once, without relying on undocumented properties
of how Timely channels behave. For operators that need to flatten the
output need slightly different logic.

The copy-to-s3 operator sees some changes, which are roughly:
* Accept the chain formed by `consolidate_pact`, and rely on it
  partitioning the data according to the exchange function.
* Skip encoding the batch, and recompute it when writing the data.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>