optimizer: Downgrade "Fixpoint detected a loop" soft_assert to error
I'd like to stop #8197 from flaking Nightly, because we have
accumulated enough examples to debug this when we get to this issue.
But at the same time, I would still be interested in if this happens
in production, because that would affect the prioritization of the
issue. (So far, it hasn't happened in production at all in the
7 months since I added this soft_panic_or_log.)