storage: Export primary ingestion collection (#30991)
Previously, when the `force_source_table_syntax` flag was enabled, the
primary source ingestion collection was not included in the source
exports. This would cause the primary source ingestion collection's
upper and since to be stuck at 0, and it would break some existing code.

This commit always includes the primary ingestion collection in the
source exports. However, when the `force_source_table_syntax` flag is
enabled, then the source export details are set to
`SourceExportDetails::None`. The result is that all source types with
the flag enabled behave similarly to how multi-output sources behave
with the flag disabled in regard to the primary ingestion collection.
Specifically, their upper's and since's move forward in time and
querying them returns an empty result.

A downside of this commit is that a source ingestion is always scheduled
with the flag enabled, even if there are no table exports. Previously,
they would only be scheduled if there were table exports.

Works towards resolving #MaterializeInc/database-issues/issues/8620